разделил регистрацию и список дел, обновил личный кабинет
