+++ ToDo_site/urls.py
@@ -1,18 +1,3 @@
-"""ToDo_site URL Configuration
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/4.1/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
 from django.contrib import admin
 from django.urls import path, include
 from todo import views
@@ -21,7 +6,7 @@ urlpatterns = [
     path('admin/', admin.site.urls),
 
     #Аутентификация
-    path('', include('todo.urls')),
+    path('', include('todo.registration.urls')),
     #Список дел
     path('', views.home, name='home'),
     path('current/', views.CurrentToDo, name='CurrentToDo'),
