+++ todo/registration/serializers.py
@@ -57,7 +57,3 @@ class ChangePasswordSerializer(serializers.Serializer):
     """
     old_password = serializers.CharField(required=True)
     new_password = serializers.CharField(required=True)
-
-
-
-
