+++ todo/templates/todo/personalaccount.html
@@ -1,3 +1,8 @@
 <h1>Личный кабинет</h1>
 <br>
+<div class="alert alert-danger" role="alert">
+    {{ user.get_username }}
+</div>
+<br>
 <a role="button" class="btn btn-primary btn-lg" href="{% url 'changepassword' %}">Cменить пароль</a>
+
